src/api-engine/requirements.txt
Django<=2.1.4
uwsgi<=2.0.17.1
djangorestframework<=3.9.0
django-filter<=1.1.0
drf-yasg<=1.11.1
flex<=6.13.2
swagger_spec_validator<=2.4.1
python-keycloak<=0.14.4
psycopg2<=2.7.6.1
holdup>1.5.0,<=1.6.0
django-extensions<=2.1.6
pygraphviz<=1.5
celery<=4.2.1
redis<=3.2.0
docker-compose>=1.18.0,<=1.23.2
watchdog<=0.9.0
